+++ sources/DDSTextureLoader.NET.csproj
