+++ sources/DDS_ALPHA_MODE.cs
