+++ sources/ManagedSubresourceData.cs
