+++ sources/TextureParsing/Caps2Flags.cs
