+++ sources/TextureParsing/D3D11_RESOURCE_DIMENSION.cs
