+++ sources/TextureParsing/D3DFormat.cs
