+++ sources/TextureParsing/DDS_MISC_FLAGS2.cs
