+++ sources/TextureParsing/DdsHeader.cs
