+++ sources/TextureParsing/DdsHeaderDxt10.cs
