+++ sources/TextureParsing/DdsPixelFormat.cs
