+++ sources/TextureParsing/FileMetadata.cs
