+++ sources/TextureParsing/HeaderFlags.cs
