+++ sources/TextureParsing/InteropTypeUtilities.cs
