+++ sources/TextureParsing/PixelFormatFlags.cs
