+++ sources/TextureParsing/Size2.cs
