+++ sources/TextureParsing/Size3.cs
