+++ sources/TextureParsing/SurfaceInfo.cs
