+++ sources/TextureParsing/ThrowHelper.cs
